Replace metric/package/recordgen w/ common extension
