+++ docs/user_guide/records.rst
@@ -7,4 +7,5 @@ To support hardware provenance, the SiliconCompiler supports automated tracking 
 
 Records are kept on a per step, and index basis. Records must be stored for each task in the flowgraph to ensure unbroken traceability from the beginning to the end in the chain of custody.
 
-.. recordgen::
+.. schema_category_summary::
+  :category: record
